requirements.txt
python-socketio==5.5.1